Update Spiral test
update Spiral test and change class name.
